refactor: Split packages
