+++ typesafe/sq/config.go
