+++ typesafe/sq/interface.go
